+++ examples/RewriterPlugin/plugin/RewriterPlugin.hs
@@ -104,7 +104,7 @@ rewrite_add pluginDefs@( PluginDefs { .. } ) _givens tys
       -> do
           wanted <- mkCancellableWanted pluginDefs b
           pure $ API.TcPluginRewriteTo
-                  ( API.mkTyFamAppReduction "RewriterPlugin" API.Nominal addTyCon tys b )
+                  ( API.mkTyFamAppReduction "RewriterPlugin" API.Nominal API.emptyVarSet addTyCon tys b )
                   [ wanted ]
       -- "Add a Zero = a", emitting a "Cancellable a" Wanted constraint.
       | Just ( zero, [] ) <- API.splitTyConApp_maybe b
@@ -112,7 +112,7 @@ rewrite_add pluginDefs@( PluginDefs { .. } ) _givens tys
       -> do
         wanted <- mkCancellableWanted pluginDefs a
         pure $ API.TcPluginRewriteTo
-                  ( API.mkTyFamAppReduction "RewriterPlugin" API.Nominal addTyCon tys a )
+                  ( API.mkTyFamAppReduction "RewriterPlugin" API.Nominal API.emptyVarSet addTyCon tys a )
                   [ wanted ]
 
       -- Erroring on 'BadNat'.
@@ -137,7 +137,7 @@ rewrite_add pluginDefs@( PluginDefs { .. } ) _givens tys
   = pure API.TcPluginNoRewrite
   where
     badRedn :: API.Reduction
-    badRedn = API.mkTyFamAppReduction "RewriterPlugin" API.Nominal
+    badRedn = API.mkTyFamAppReduction "RewriterPlugin" API.emptyVarSet API.Nominal
       addTyCon tys (API.mkTyConApp badNatTyCon [])
 
 -- Given the type "a", constructs a "Cancellable a" constraint